rename folder to bench
