+++ benchmark/bench/index.js
