+++ benchmark/bench/src/ascii-chart.js
