+++ benchmark/bench/src/histograms.js
