+++ benchmark/bench/src/performance-entry-views.js
